Add support for SUBIFD and EXTRASAMPLES tags with ctypes wrapper
These tags are special as they are variable length, so they are stored as a count + pointer.
So we introduce a special way to encode such tags.
This also means the "count" argument of SetField() is actually not needed anymore,
as it's all done automatically.

The SUBIFD tag allows to build "TIFF trees".
The libtiff convention is that you first write this tag an array of length N on a directory,
and then the next N directories written will be in this sub-directory.

The EXTRASAMPLE tag allows to handle unsual TIFF files with multiple extra-samples (eg, several masks).