Do not even attempt to read COLORMAP tag if bits-per-sample > 16
When copying tags, the copy() function looks for every potential tag.
It causes with COLORMAP when bits-per-sample is large, as to look for the tag,
the only is to first allocate the memory needed in case the tag is present.
Ideally, libtiff would provide a function to check whether the tag is present,
without reading it. Internally, there is TIFFFieldSet(), but it's not exported.

=> Don't even try to read the tag with such high bps, and instead report the
tag is not present.