ctypes wrapper: fix opening files with names containing non-ascii characters
The changes to support Python 3 forced the filename to be explictily encoded.
However, the encoding was hard-coded to "ascii", which prevent any file with
non-ascii character to be opened.
Python 3 offers a much better function to use: os.fsencode()
Unfortunately, Python 2 doesn't have this function, but it's possible to
recreate a simple version of it when unicode is passed.