Add header for libtiff 4.5.0.
Also, there was a formatting change between 4.4.0 and 4.5.0 where some
definitions are split across multiple lines.  This adds some code to
merge those lines together so that it the parsing still works.

Perhaps uninterestingly, the result for 4.5.0 is the same as 4.4.0 (and
4.3.0).

Lastly, look in /usr/local/include for the header file as well as the
other locations.