libtiff_ctype: On Windows, use TIFFOpenW() when open() gets a str
Depending the Windows encoding, it can be impossible to encode a unicode
path into the "filesystem encoding". So instead of employing all sorts
of hacks, just use the designated function which directly takes unicode
input. This has the advantage of being much more reliable.

Also drop Python 2 support, as it's no needed anymore.