ctypes: don't fail to load when embedded as a "frozen" package
When the module is embedded in a executable (aka "frozen"), the __file__
path is virtual, and might not actually exists on the computer.
This path is used at init to look for the DLL, by changing the current
working directory to the containing folder. As it doesn't exist, it
fails, and libtiff_ctypes cannot be imported.

Instead of completely failing, now handle such case, and just don't use
that folder as an extra location to look for the DLL.