ctypes: don't fail to load when embedded as a "frozen" package (#187)
* ctypes: don't fail to load when embedded as a "frozen" package

When the module is embedded in a executable (aka "frozen"), the __file__
path is virtual, and might not actually exists on the computer.
This path is used at init to look for the DLL, by changing the current
working directory to the containing folder. As it doesn't exist, it
fails, and libtiff_ctypes cannot be imported.

Instead of completely failing, now handle such case, and just don't use
that folder as an extra location to look for the DLL.

* lsm: drop useless "global" usage

As noted by the latest flake8:
F824 `global tiff_module_dict` is unused: name is never assigned in scope

As the tiff_module_dict variable is only read, there is no need to use
a "global".